Solve format issue with real data

+++ src/components/timeChart.jsx
@@ -37,10 +37,10 @@ function TimeChart(props) {
             let hours = entry["hours"];
 
             if (!(label in lookUp)) {
-                lookUp[label] = parseInt(hours);
+                lookUp[label] = parseFloat(hours);
                 labels.push(label);
             } else {
-                lookUp[label] += parseInt(hours);
+                lookUp[label] += parseFloat(hours);
             }
         })
